v3.1
- Bug Fixes
  - Fixed database crash bug
  - Fixed unknown interaction bug
- Updated permissions

+++ Commands/PingCommand.js
@@ -1,4 +1,5 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
+const { PermissionFlagsBits } = require('discord.js');
 
 module.exports = {
 	data: new SlashCommandBuilder()
@@ -8,6 +9,12 @@ module.exports = {
         if(!interaction.guild) return
 		if(interaction.content==undefined){
             //Interaction
+			//Permission Check
+			if(!interaction.channel.permissionsFor(client.user).has(PermissionFlagsBits.SendMessages) || !interaction.channel.permissionsFor(client.user).has(PermissionFlagsBits.ViewChannel)){
+				return interaction.reply({ content: `I'm sorry, I do not have enough permissions for this command! \n I need \`Send Messages\`, and \`Read Message History\``, ephemeral: true }).catch(() => {
+					return; 
+				})
+			}
             const m = await interaction.channel.send("Ping?");
             await interaction.reply(`:ping_pong: Pong! | Latency is \`${m.createdTimestamp - interaction.createdTimestamp}ms\` | API latency is \`${client.ws.ping}ms\``)
             m.delete()
